Incorporated user multitenancy with one passing spec.
The feature test passes, but is not robust as the view simply displays the name. In order to satisfy the client, we should populate that view soon. The multitenancy was very easy, but it's effects were widespread. Fixing those issues and passing slug params around will be the big challenge of this first iteration.