Fixed most broken specs.
I messed around a bit with the namespacing/multitenacity, but it seems more complicated than I anticipated. We'll have to tread carefully with that. Also, the cats and breeds will have to be renamed and the tables remapped for them. That will break every single test, but it shouldn't be hard to replace them.