Fixed the cart so the view is clean.
The first issue was passing in an array of arrays which sloppified the view with index calls. Almost every method in the cart model reffered to the cat_quantity array. We decided to switch it to a hash which was not only simpler to work with in the model, but also easier to call in the view. Lastly we made a 'subtotal' method which solved the only bit of math in the view. This commit also added the brakeman gem which gives us 0 security flaws regardless.