Added validation flash messages to login modal functionality.
We first linked the User.create.errors to the application.html. Then we created a new conditional after that to display an extra flash message if the login fails. This prompts the user to try logging in again via the modal. The next step is implementing this with the signup modal and making sure the only flash messages which appear are appropriate.