Further refactored the validations and flash messages.
Firstly, I edited the logo image a bit, but it still isn't perfect. Give me another 10 hours on photoshop. Dealing with the validations and modal, I found it hard to  and pass in something for the view to handle. The solution is to only have a link for the login modal, and not even worry about the signin modal.